backend_django/requirements.txt
Django==4.2.16
djangorestframework==3.15.2
django-cors-headers==4.4.0
bcrypt==4.2.1
python-jose==3.3.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
